added comment in cleardatepresparkrequesthook ClearDataPreSparkRequestHook() method made changes

+++ thingifier/src/main/java/uk/co/compendiumdev/thingifier/application/sparkhttpmessageHooks/ClearDataPreSparkRequestHook.java
@@ -9,11 +9,10 @@ public class ClearDataPreSparkRequestHook implements SparkRequestResponseHook {
     private final Thingifier thingifier;
     private final long maxgap;
 
-    //TODO: cast int minutes to long
     public ClearDataPreSparkRequestHook(final int minutes, Thingifier thingifier) {
         this.lastReset = System.currentTimeMillis();
         this.thingifier = thingifier;
-        this.maxgap = minutes*60*1000;
+        this.maxgap = (long) minutes *60*1000;
     }
 
     @Override
